Update: Create classes and text files
Create User class which can be inherited by both
Customer and Admin classes. User class contains
methods that both Customer and Admin can use, e.g.:
    1. Login
    2. Logout
    3. Save personal data to a text file

Create text files that will act as our database
for this project. Text files are stored in
Database folder in the Java project. A saveData
method is added in the User class for the Customer
and Admin classes to implement to save data to their
respective text files.